Remove fonts
I am moving away from maintaing static assets from within the repository. It
isn't practical and just unecessarily increases the size of your repository,
From now on I will be using CDNs or simillar solutions to deliver static
content. This decision is based on my particular experience and within the scope
of this theme and my personal blog which uses this theme. For a different
project I might do the oppsite.